+++ _boards/evkcmimxrt1060/ota_examples/ota_mcuboot_client/enet/app.h
@@ -10,7 +10,6 @@
 
 /*${header:start}*/
 #include "fsl_phyksz8081.h"
-#include "ksdk_mbedtls.h"
 /*${header:end}*/
 
 /*******************************************************************************