+++ _boards/evkcmimxrt1060/ota_examples/ota_mcuboot_client/mbedtls_user_config.h
@@ -8,13 +8,8 @@
 #ifndef MBEDTLS_USER_CONFIG_H
 #define MBEDTLS_USER_CONFIG_H
 
-// on some RT targets, mainly RT10xx, the following is needed to specify the ciphersuit manually
-// to avoid incompatible setup that is not supported by HW acceleration
-
-#if defined(MBEDTLS_FREESCALE_DCP_AES) && defined(MBEDTLS_AES_ALT_NO_256)
-#define MBEDTLS_SSL_CIPHERSUITES                                                                    \
-    MBEDTLS_TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256, MBEDTLS_TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256, \
-        MBEDTLS_TLS_RSA_WITH_AES_128_GCM_SHA256
-#endif
+#define MBEDTLS_PLATFORM_MEMORY
+#define MBEDTLS_PLATFORM_STD_CALLOC        pvPortCalloc
+#define MBEDTLS_PLATFORM_STD_FREE          vPortFree
 
 #endif /* MBEDTLS_USER_CONFIG_H */