+++ _boards/mcxn5xxevk/ota_examples/ota_linker/linker_reconfig.cmake
@@ -11,11 +11,5 @@ mcux_remove_armgcc_linker_script(
     LINKER ${device_root}/${soc_portfolio}/${soc_series}/${device}/gcc/${CONFIG_MCUX_TOOLCHAIN_LINKER_DEVICE_PREFIX}_flash.ld
 )
 
-mcux_remove_mdk_linker_script(
-    BASE_PATH ${SdkRootDirPath}
-    TARGETS debug release
-    LINKER ${device_root}/${soc_portfolio}/${soc_series}/${device}/arm/${CONFIG_MCUX_TOOLCHAIN_LINKER_DEVICE_PREFIX}_flash.scf
-)
-
 # Specific linkers are added using project segments
 # see examples/_common/project_segments/ota_examples/mcxn_boards/prjseg.cmake