+++ ota_examples/ota_mcuboot_client/FreeRTOSConfig.h
@@ -72,7 +72,7 @@
 #define configAPPLICATION_ALLOCATED_HEAP        0
 
 #ifndef configTOTAL_HEAP_SIZE
-#define configTOTAL_HEAP_SIZE                   ( ( size_t ) ( 100 * 1024 ) )    /* 100kB. */
+#define configTOTAL_HEAP_SIZE                   ( ( size_t ) ( 128 * 1024 ) )    /* 128kB. */
 #endif